Add catalog entities, FixtureProvider port, seed provider, sync service
Competition/Team/Match map the V1 schema (match scores widened to integer
to match entity types). FixtureProvider is the provider-neutral port from
the design doc; the seed adapter generates deterministic fixtures relative
to the clock (finished / in-play / upcoming at +2h,+1d,+2d,+3d) so the app
runs offline. FixtureSyncService upserts by provider_ref and is verified
idempotent. Dev profile bootstraps an admin account and a seeded catalog.

+++ backend/src/main/resources/db/migration/V1__core_schema.sql
@@ -25,8 +25,8 @@ create table matches (
     away_team_id   bigint      not null references teams (id),
     kickoff_utc    timestamptz not null,
     status         varchar(16) not null default 'SCHEDULED',
-    home_score     smallint,
-    away_score     smallint,
+    home_score     integer,
+    away_score     integer,
     provider_ref   varchar(64) unique,
     last_synced_at timestamptz,
     constraint chk_matches_teams_differ check (home_team_id <> away_team_id)
@@ -69,9 +69,9 @@ create table predictions (
     id                  bigint generated always as identity primary key,
     user_id             bigint      not null references users (id),
     gameweek_fixture_id bigint      not null references gameweek_fixtures (id) on delete cascade,
-    home_goals          smallint    not null check (home_goals between 0 and 20),
-    away_goals          smallint    not null check (away_goals between 0 and 20),
-    points              smallint, -- null until the match is scored
+    home_goals          integer     not null check (home_goals between 0 and 20),
+    away_goals          integer     not null check (away_goals between 0 and 20),
+    points              integer, -- null until the match is scored
     scored_at           timestamptz,
     updated_at          timestamptz not null default now(),
     unique (user_id, gameweek_fixture_id)
